Initial implementation
The resourceLoop() function is based on infrastructure cloned from kubernetes-aws-resource-service, and then adapted
to TypeScript.

Compared to the initial sketch in the README.md some things had to be changed, in particular this implementation avoids
querying the deployments/statefulsets/... for each observed update, as that got _extremely_ expensive (to the point of
harming the cluster/kubelet/api-server? stability). The problem here is actually that Kubernetes/EKS doesn't clean out
secrets for service accounts, even when that particular secret is no longer referenced. At this point we have _thousands_
of such secrets in the master environment alone.

Note that the auto-kubernetes-client related types should move there eventually.